% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_minimum_interval_ms}
\alias{get_minimum_interval_ms}
\title{Get Minimum Interval in Milliseconds}
\usage{
get_minimum_interval_ms(handle)
}
\arguments{
\item{handle}{Integer handle of unit}
}
\description{
Get Minimum Interval in Milliseconds
}
\details{
Can answer zero.
}
