% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_single}
\alias{get_single}
\title{Get Single Temperature}
\usage{
get_single(handle, units)
}
\description{
Get Single Temperature
}
\details{
Uses \code{USBTC08_CHANNEL_COUNT} rather than \code{USBTC08_MAX_CHANNELS} because each unit has nine channels including CJC.
}
