% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_temp}
\alias{get_temp}
\title{Get Temperature}
\usage{
get_temp(handle, buffer_length, channel, units, fill_missing)
}
\description{
Get Temperature
}
\details{
The number of readings returned may not match the length requested, or might even be zero.
}
