% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_unit_info}
\alias{get_unit_info}
\title{Get Unit Information}
\usage{
get_unit_info(handle)
}
\arguments{
\item{handle}{Integer handle of unit}
}
\description{
Get Unit Information
}
\details{
Makes an important assumption about the driver version: the driver
must null-terminate the character string. Also assumes that non-zero
return indicates success and never error.
}
