% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_channel}
\alias{set_channel}
\title{Set Channel Type}
\usage{
set_channel(handle, channel, tc_type)
}
\arguments{
\item{handle}{Unit handle}

\item{channel}{Selected channel}

\item{tc_type}{Thermocouple type, one of the follow
\code{c("B", "E", "J", "K", "N", "R", "S", "T")}
or \code{"X"} for voltage, space \code{" "} to disable channel}
}
\description{
Set Channel Type
}
