% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{usb_tc08_get_last_error}
\alias{usb_tc08_get_last_error}
\title{Gets last error of TC-08 unit}
\usage{
usb_tc08_get_last_error(x)
}
\arguments{
\item{x}{Open TC-08 unit.}
}
\description{
Gets last error of TC-08 unit
}
