% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{usb_tc08_get_temp}
\alias{usb_tc08_get_temp}
\title{Get temperatures asynchronously}
\usage{
usb_tc08_get_temp(x)
}
\arguments{
\item{x}{Opened and running TC-08 in asynchronous mode.}
}
\description{
Get temperatures asynchronously
}
