% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{usb_tc08_get_temp_deskew}
\alias{usb_tc08_get_temp_deskew}
\title{Get asynchronously-captured de-skewed temperatures}
\usage{
usb_tc08_get_temp_deskew(x)
}
\arguments{
\item{x}{Opened and running TC-08 in asynchronous mode.}
}
\description{
Get asynchronously-captured de-skewed temperatures
}
