% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{usb_tc08_open}
\alias{usb_tc08_open}
\title{Opens TC-08 unit for synchronous-mode operations}
\usage{
usb_tc08_open()
}
\value{
Handle of opened TC-08 unit or \code{NULL} if no more closed units
available for opening.
}
\description{
Opens the next available closed TC-08 unit. Opens synchronously
and therefore blocks the caller until a USB-connected TC-08 has been found
and opened; a process that might take a second or two.
}
\details{
Asks for the last error using a zero unit handle if opening fails.
Handles only represent open units so no handle exists for closed units.
}
\examples{
\donttest{
tryCatch(usb_tc08_open(), error = function(e) e)
}
}
