% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{usb_tc08_open_async_complete}
\alias{usb_tc08_open_async_complete}
\title{Opens all TC-08 units asynchronously to completion.}
\usage{
usb_tc08_open_async_complete()
}
\value{
List of TC-08 units opened and ready for set up and running.
}
\description{
Opens all TC-08 units asynchronously to completion.
}
\details{
Runs aggressively with \emph{no} delay between pending progress
results.
}
\examples{
\donttest{
library(usbtc08r)
usb_tc08_open_async()
tc08 <- usb_tc08_open_async_complete()
lapply(tc08, usb_tc08_set_channel, 1L, "K")
lapply(tc08, usb_tc08_run, 500L)
}
}
