% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{usb_tc08_set_channel}
\alias{usb_tc08_set_channel}
\title{Sets a channel thermocouple type}
\usage{
usb_tc08_set_channel(x, ...)
}
\arguments{
\item{x}{TC-08 unit, synchronous or asynchronous mode.}

\item{channel}{Channel to set.}

\item{tc_type}{Type of thermocouple. Must be an integer, not a character.}
}
\description{
Sets a channel thermocouple type
}
