% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{usb_tc08_stop}
\alias{usb_tc08_stop}
\title{Stops an asynchronous TC-08.}
\usage{
usb_tc08_stop(x)
}
\arguments{
\item{x}{Opened and running TC-08 in asynchronous mode.}
}
\description{
Stops an asynchronous TC-08.
}
