% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{usb_tc08_open_progress}
\alias{usb_tc08_open_progress}
\title{Progressively opens TC-08 units for asynchronous-mode operations}
\usage{
usb_tc08_open_progress()
}
\description{
Progressively opens TC-08 units for asynchronous-mode operations
}
