% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usbtc08r-package.R
\docType{package}
\name{usbtc08r-package}
\alias{usbtc08r}
\alias{usbtc08r-package}
\title{usbtc08r: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Roy Ratcliffe \email{roy@ratcliffe.me} (\href{https://orcid.org/0000-0001-6594-1280}{ORCID})

}
\keyword{internal}
